+++ src/main/java/roomescape/config/LoginMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.controller.member;
+package roomescape.config;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
