+++ src/main/java/roomescape/config/WebMvcConfig.java
@@ -4,8 +4,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.controller.CheckAdminInterceptor;
-import roomescape.controller.member.LoginMemberArgumentResolver;
 import roomescape.service.MemberService;
 
 import java.util.List;
